+++ src/components/content/diagnostic-reports/helpers/diagnostic-report-query-filters.ts
@@ -1,12 +1,15 @@
-import { DiagnosticReportModel } from "@/fhir/models";
+import { DiagnosticReportModel, ObservationModel } from "@/fhir/models";
 import { ResourceMap } from "@/fhir/types";
 import { isEqual, uniqWith } from "@/utils/nodash";
 
 export const applyDiagnosticReportFilters = (
   data: fhir4.DiagnosticReport[],
-  includedResources?: ResourceMap
+  includedResources?: ResourceMap,
+  trends?: ObservationModel[]
 ) => {
-  const diagnosticModel = data.map((dr) => new DiagnosticReportModel(dr, includedResources));
+  const diagnosticModel = data.map(
+    (dr) => new DiagnosticReportModel(dr, includedResources, undefined, trends)
+  );
   const diagnosticReportData = uniqWith(diagnosticModel, (a, b) =>
     isEqual(valuesToDedupeOn(a), valuesToDedupeOn(b))
   );