+++ src/components/content/timeline/helpers/filters.tsx
@@ -1,5 +1,4 @@
 import { faClipboardCheck } from "@fortawesome/free-solid-svg-icons";
-import { BetaLabel } from "@/components/core/beta-label";
 import { FilterChangeEvent, FilterItem } from "@/components/core/filter-bar/filter-bar-types";
 import { TimelineEventModel } from "@/fhir/models/timeline-event";
 import { compact, isEqual, uniqWith } from "@/utils/nodash/fp";
@@ -20,11 +19,6 @@ export function timelineFilters(timelineEvents: TimelineEventModel[]): FilterIte
             ? {
                 key: te.type,
                 name: te.type,
-                display: te.beta ? (
-                  <span>
-                    {te.type} <BetaLabel />
-                  </span>
-                ) : undefined,
               }
             : undefined
         )